Test: report test phases
The avocado test runner mostly "flies blind" when it comes to the
individual test phases, that is, it doesn't know if a test is
currently being initialized, running its setup, running the test
method itself or its teardown.

With this, the runner gets access to that information, and can act
differently based on this knowledge.  One use case is to allow for
different timeouts on different test phases.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>